:construction: Canvas3 Class in progress
