+++ components/canvas/Image.vue
@@ -13,15 +13,15 @@
 </template>
 
 <script setup>
-// import { Canvas } from "~/utils/canvas";
-import { useCanvas3Store } from "~/stores/canvas3";
+import { Canvas3 } from "~/utils/canvas3";
+// import { useCanvas3Store } from "~/stores/canvas3";
 
 // import { useDisplayStore } from "~/stores/display";
 
 // const displayStore = useDisplayStore();
 // const navigationStore = useNavigationStore();
 
-const Canvas3 = useCanvas3Store();
+// const Canvas3 = useCanvas3Store();
 
 const props = defineProps({
   alt: {
