+++ components/canvas/Text.vue
@@ -9,15 +9,16 @@
 </template>
 
 <script setup>
-import { useCanvas3Store } from "~/stores/canvas3";
+// import { useCanvas3Store } from "~/stores/canvas3";
 
-// import { Canvas } from "~/utils/canvas";
+import { Canvas3 } from "~/utils/canvas3";
+
 // import { useDisplayStore } from "~/stores/display";
 
 // const displayStore = useDisplayStore();
 // const navigationStore = useNavigationStore();
 
-const Canvas3 = useCanvas3Store();
+// const Canvas3 = useCanvas3Store();
 
 const props = defineProps({
   shader: {
